+++ src/query/service/tests/it/pipelines/mod.rs
@@ -13,4 +13,3 @@
 // limitations under the License.
 
 mod executor;
-mod processors;
